Move flow-analysis parse option to base test helpers
