+++ src/Test.Utilities/CSharpCodeFixVerifier`2+Test.cs
@@ -1,5 +1,8 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Testing;
@@ -25,6 +28,17 @@ namespace Test.Utilities
 
                     return solution;
                 });
+            }
+
+            protected override Project ApplyCompilationOptions(Project project)
+            {
+                var newProject = base.ApplyCompilationOptions(project);
+
+                var parseOptions = newProject.ParseOptions.WithFeatures(
+                    newProject.ParseOptions.Features.Concat(
+                        new[] { new KeyValuePair<string, string>("flow-analysis", "true") }));
+
+                return newProject.WithParseOptions(parseOptions);
             }
 
             public LanguageVersion LanguageVersion { get; set; } = LanguageVersion.CSharp7_3;
