+++ src/Test.Utilities/VisualBasicSecurityCodeFixVerifier`2+Test.cs
@@ -1,8 +1,5 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Testing;
@@ -19,17 +16,6 @@ namespace Test.Utilities
             {
                 // These analyzers run on generated code by default.
                 TestBehaviors |= TestBehaviors.SkipGeneratedCodeCheck;
-            }
-
-            protected override Project ApplyCompilationOptions(Project project)
-            {
-                var newProject = base.ApplyCompilationOptions(project);
-
-                var parseOptions = newProject.ParseOptions.WithFeatures(
-                    newProject.ParseOptions.Features.Concat(
-                        new[] { new KeyValuePair<string, string>("flow-analysis", "true") }));
-
-                return newProject.WithParseOptions(parseOptions);
             }
         }
     }
